+++ src/main/java/maven/fetcher/internal/MavenDependencyFetcher.java
@@ -25,6 +25,7 @@ public class MavenDependencyFetcher implements DependencySelector {
     private final Collection<String> scopes;
     private final boolean retrieveOptionals;
     private final List<Dependency> dependencies;
+    private final List<Exclusion> exclusions;
     private final List<RemoteRepository> remoteRepositories;
     private final RepositorySystem system;
     private final Logger logger;
@@ -44,9 +45,12 @@ public class MavenDependencyFetcher implements DependencySelector {
         this.session = session.setDependencySelector(this);
         this.scopes = fetchRequest.scopes();
         this.retrieveOptionals = fetchRequest.isRetrievingOptionals();
+        this.exclusions = fetchRequest.excludedArtifacts().stream()
+            .map(this::exclusionFromCoordinates)
+            .collect(Collectors.toList());
         this.dependencies = fetchRequest.artifacts().stream()
             .map(this::artifactFromCoordinates)                        
-            .map(artifact -> new Dependency(artifact, null))
+            .map(artifact -> new Dependency(artifact, null, false, exclusions))
             .collect(Collectors.toList());
         this.logger = logger;
     }
@@ -60,7 +64,7 @@ public class MavenDependencyFetcher implements DependencySelector {
             }
         }
         this.retrievedArtifacts = new HashSet<>();
-        CollectRequest request = new CollectRequest(dependencies, null, remoteRepositories);
+        CollectRequest request = new CollectRequest(dependencies, dependencies, remoteRepositories);
         CollectResult result = system.collectDependencies(session, request);
         retrieveDependency(result.getRoot());
         return new MavenFetchResultImpl(result, session);
@@ -90,8 +94,25 @@ public class MavenDependencyFetcher implements DependencySelector {
         }
     }
 
+
+    private Exclusion exclusionFromCoordinates (String coordinates) {
+        var parts = coordinates.split(":");
+        if (parts.length >= 2) {
+            var groupId = parts[0];
+            var artifactId = parts[1];
+            return new Exclusion(groupId, artifactId, "jar", null);
+        } else {
+            throw new IllegalArgumentException("Invalid exclusion '"+coordinates+"'");
+        }
+    }
+
+
+
     private void retrieveDependency(DependencyNode node) {
         if (node.getArtifact() != null) {
+           if (isExcluded(node.getArtifact())) {
+               return;
+           }
             try {
                 ArtifactRequest request = new ArtifactRequest(
                     node.getArtifact(), remoteRepositories, null
@@ -106,6 +127,18 @@ public class MavenDependencyFetcher implements DependencySelector {
         for (DependencyNode child : node.getChildren()) {
             retrieveDependency(child);
         }
+    }
+
+
+    private boolean isExcluded(Artifact artifact) {
+        for (Exclusion exclusion : exclusions) {
+            if (exclusion.getGroupId().equals(artifact.getGroupId()) &&
+                exclusion.getArtifactId().equals(artifact.getArtifactId())
+            ) {
+                return true;
+            }
+        }
+        return false;
     }